requirements.txt
psycopg[binary]
PyYAML
psycopg_pool 
python-dotenv
pandas